Ajuste tabela endereço

+++ src/main/java/com/dasad/empresa/controller/EnderecoController.java
@@ -25,20 +25,29 @@ public class EnderecoController implements EnderecoApi {
     public EnderecoController() {
     }
 
+    @Override
     @PutMapping({"/{id}"})
-    public ResponseEntity<EnderecoModel> updateEndereco(@PathVariable Integer id, @RequestBody EnderecoModel enderecoDetails) {
+    public ResponseEntity<EnderecoModel> updateendereco(@PathVariable Integer id, @RequestBody EnderecoModel enderecoDetails) {
         return this.enderecoService.findById(id).map((endereco) -> {
             endereco.setLogradouro(enderecoDetails.getLogradouro());
-            endereco.setCidade(enderecoDetails.getCidade());
-//            endereco.setEstado(enderecoDetails.getEstado());
+            endereco.setCidadeId(enderecoDetails.getCidadeId());
             endereco.setCep(enderecoDetails.getCep());
+            endereco.setNumero(enderecoDetails.getNumero());
+            endereco.setComplemento(enderecoDetails.getComplemento());
+            endereco.setBairro(enderecoDetails.getBairro());
             EnderecoModel enderecoSave = this.enderecoService.save(endereco);
+
             return ResponseEntity.ok(enderecoSave);
         }).orElse(ResponseEntity.notFound().build());
     }
 
+    @Override
     @DeleteMapping({"/{id}"})
-    public ResponseEntity<Void> deleteEndereco(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteendereco(@PathVariable Integer id) {
+        if (id == null) {
+            return ResponseEntity.badRequest().build();
+        }
+
         if (this.enderecoService.findById(id).isPresent()) {
             this.enderecoService.deleteById(id);
             return ResponseEntity.noContent().build();
@@ -49,20 +58,28 @@ public class EnderecoController implements EnderecoApi {
 
     @Override
     @PostMapping
-    public ResponseEntity<EnderecoModel> createEndereco(@RequestBody EnderecoModel enderecoModel) {
+    public ResponseEntity<EnderecoModel> createendereco(@RequestBody EnderecoModel enderecoModel) {
+        if (enderecoModel == null) {
+            return ResponseEntity.badRequest().build();
+        }
+
         var endereco = this.enderecoService.save(enderecoModel);
         return ResponseEntity.ok(endereco);
     }
 
     @Override
     @GetMapping({"/{id}"})
-    public ResponseEntity<EnderecoModel> detailEndereco(Integer id) {
+    public ResponseEntity<EnderecoModel> detailendereco(Integer id) {
+        if (id == null) {
+            return ResponseEntity.badRequest().build();
+        }
+
         return this.enderecoService.findById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
 
     @Override
     @GetMapping
-    public ResponseEntity<List<EnderecoModel>> findEndereco() {
+    public ResponseEntity<List<EnderecoModel>> findendereco() {
         var enderecos =  this.enderecoService.findAll();
         return ResponseEntity.ok(enderecos);
     }
